Поиск по имени через foreach Тестовый список

+++ src/main/java/ru/ugrasu/journal/controllers/UserController.java
@@ -23,15 +23,34 @@ public class UserController {
     @Autowired
     private UserService userService;
 
-    //TODO
-    //1. Обработка Null-value
-    //2. Обработка исключений
+    //TODO Общий список недоработок:
+    //1. Закончить структуру БД
+    //2. Написать вьюшку SQL для удобства обработки
+    //3. Расписать функционал
+    //4. Обработка Null-value
+    //5. Обработка исключений
+    //6. Связи между таблицами
+    //7. Замена значений в связях
+    //8. Можно возвращать ResponseEntity (вернет так же указанный статус HTTP параметром)
+    //9. Frontend нарисовать)
+    //10. При редактировании пользователя логичнее сделать выпадающий список
 
     @RequestMapping(value = "/findById/{id}", produces = APPLICATION_JSON_UTF8_VALUE, method = GET)
     public UserEntity findById(@PathVariable("id") int id) {
         System.out.println("Execution: controller - findById - " + id);
+        UserEntity userEntity = userService.findById(id);
+        if (userEntity == null) {
+            throw new ResourceNotFoundException("Teacher not found");
+        } else {
+            return userEntity;
+        }
+    }
+
+    @RequestMapping(value = "/findByName/{name}", produces = APPLICATION_JSON_UTF8_VALUE, method = GET)
+    public List<UserEntity> findByName(@PathVariable("name") String name) {
+        System.out.println("Execution: controller - findByName - " + name);
         try {
-            return userService.findById(id);
+            return userService.findByName(name);
         } catch (Exception e) {
             throw new ResourceNotFoundException("Teacher not found");
         }
@@ -40,13 +59,20 @@ public class UserController {
     @RequestMapping(value = "/save", consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE, method = RequestMethod.POST)
     public UserEntity save(UserEntity newUser){
         return userService.save(newUser);
+
         //TODO Response возвращает со статусом
         //return new ResponseEntity<UserEntity>(newUser, HttpStatus.OK);
     }
 
     @RequestMapping(value = "/deleteById/{id}", produces = APPLICATION_JSON_UTF8_VALUE, method = GET)
     public void delete (@PathVariable("id") Integer id) {
-        userService.deleteById(id);
+        System.out.println("Execution: controller - deleteById - " + id);
+        UserEntity userEntity = findById(id);
+        if (userEntity == null) {
+            throw new ResourceNotFoundException("Teacher not found");
+        } else {
+            userService.deleteById(id);
+        }
     }
 
 }
